separated components into individual subfolders for easier navigation
