+++ client/src/components/TeamBuilder/SearchBar/SearchBar.tsx
